Extract response into multiple classes
